Add an action for linting and releasing helm chart
I've added a GitHub action that is linting helm chart on each pushed
commit and releasing a new version helm chart automatically on each new
push to the `master` branch.

The linter job is checking whether the chart is changed, and in case it
is, the linter expects the versions to be bumped. That means that all
the changes to chart will be delivered to its users right away.